requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-google-genai>=2.0.0
langsmith>=0.1.0
google-generativeai>=0.8.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.28.0
jupyter
ipykernel